3-D object created on chessboard

+++ src/calibrate_camera.cpp
@@ -1,10 +1,10 @@
 /**
- * @file vidDisplay.cpp
+ * @file calibrate_camera.cpp
  * @author Samreen
- * @brief starts video, and calibrates the camera
- * different keypresses
+ * @brief starts the video stream and calibrates the camera and saves the
+ * parameters in a csv file.
  * @version 0.1
- * @date 2022-01-28
+ * @date 2022-03-27
  *
  * @copyright Copyright (c) 2022
  *
@@ -20,7 +20,7 @@
 using namespace std;
 using namespace cv;
 
-int cornersExist = 1;
+bool cornersExist = false;
 vector<Point2f> lastcorners;
 vector<vector<Point2f>> allcorners;
 vector<Vec3f> points;
@@ -46,7 +46,11 @@ int main(int argc, char *argv[]) {
   Mat frame;
 
   // get the points for the chessboard
-  getPoints(points);
+  getPoints(points, 0);
+
+  Mat cameraMatrix = (Mat_<float>(3, 3) << 1, 0, frame.cols / 2, 0, 1,
+                      frame.rows / 2, 0, 0, 1);
+  Mat distCoeffs, rvecs, tvecs;
 
   for (;;) {
     *capdev >> frame;  // get a new frame from the camera, treat as a stream
@@ -66,11 +70,11 @@ int main(int argc, char *argv[]) {
     // get the last corners
     cornersExist = getLastCorners(frame, lastcorners);
 
-    if (key == 's') {
-      // print lasr corners
-      // cout << "last corners: " << lastcorners << endl;
-
+    if (key == 's' && cornersExist) {
+      // add corner to list of corners
       allcorners.push_back(vector<Point2f>(lastcorners));
+      // print number of frames saved
+      printf("Saving frame %d\n", allcorners.size());
       drawChessboardCorners(frame, Size(9, 6), lastcorners, true);
       string path = "data/image_" + to_string(allcorners.size()) + ".jpg ";
       imwrite(path, frame);
@@ -81,19 +85,22 @@ int main(int argc, char *argv[]) {
       // callibrate the camera if at least 5 calibration frames exist
       if (allcorners.size() > 5) {
         cout << "Calibrating the camera..." << endl;
-        Mat cameraMatrix = (Mat_<float>(3, 3) << 1, 0, frame.cols / 2, 0, 1,
-                            frame.rows / 2, 0, 0, 1);
-        Mat distCoeffs, rvecs, tvecs;
-        RPerror =
-            calibrateCamera(allpoints, allcorners, refS, cameraMatrix,
-                            distCoeffs, rvecs, tvecs, CALIB_FIX_ASPECT_RATIO);
-        cout << "camera matrix: " << cameraMatrix << endl;
-        cout << "distortion coefficients: " << distCoeffs << endl;
-        cout << "reprojection error: " << RPerror << endl;
+
+        RPerror = calibrateCamera(allpoints, allcorners, refS, cameraMatrix,
+                                  distCoeffs, rvecs, tvecs, 0);
+        // cout << "Camera matrix: " << cameraMatrix << endl;
+        // cout << "Distortion coefficients: " << distCoeffs << endl;
+        cout << "Reprojection error: " << RPerror << endl;
+        // cout << "Rvecs: " << rvecs << endl;
+        // cout << "Tvecs: " << tvecs << endl;
       }
+    } else if (key == 'w' && allcorners.size() > 5) {
+      // write the parameters to a csv file
+      cout << "Writing parameters to csv file..." << endl;
+      writeParametersToCSV(cameraMatrix, distCoeffs, rvecs, tvecs);
     }
 
-    if (cornersExist == 0) {
+    if (cornersExist) {
       drawChessboardCorners(frame, Size(9, 6), lastcorners, true);
     }
